setup TopNav to not render the week dropdown unless the week is passed in. Rename StatsProvider to StatsPage because it has layout responsibilty as well
